Enable management file
