+++ test/e2e/boyar_driver.go
@@ -1,23 +1,13 @@
 package e2e
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/docker/docker/api/types/filters"
-	"github.com/orbs-network/boyarin/boyar/config"
-	"github.com/orbs-network/boyarin/services"
-	"github.com/orbs-network/boyarin/strelets/adapter"
 	"github.com/orbs-network/boyarin/test/helpers"
-	"github.com/orbs-network/govnr"
-	"github.com/orbs-network/scribe/log"
 	"github.com/stretchr/testify/require"
 	"io/ioutil"
-	"net"
 	"net/http"
-	"net/http/httptest"
 	"os"
-	"strconv"
 	"testing"
 	"time"
 )
@@ -41,10 +31,10 @@ func (vc VChainArgument) ExternalPort() int {
 	return port + vc.Id
 }
 
-func configJson(t *testing.T, topology []interface{}, genesisValidators []string, httpPort int, vChains []VChainArgument) string {
+func configJson(t *testing.T, topology []interface{}, managementUrl string, httpPort int, vChains []VChainArgument) string {
 	chains := make([]interface{}, len(vChains))
 	model := map[string]interface{}{
-		"network": topology,
+		"network": []string{},
 		"orchestrator": map[string]interface{}{
 			"max-reload-time-delay": "1s",
 			"http-port":             httpPort,
@@ -67,14 +57,14 @@ func configJson(t *testing.T, topology []interface{}, genesisValidators []string
 		},
 	}
 	for i, id := range vChains {
-		chains[i] = VChainConfig(id, genesisValidators)
+		chains[i] = VChainConfig(id, managementUrl)
 	}
 	jsonStr, err := json.MarshalIndent(model, "", "    ")
 	require.NoError(t, err)
 	return string(jsonStr)
 }
 
-func VChainConfig(vc VChainArgument, genesisValidators []string) map[string]interface{} {
+func VChainConfig(vc VChainArgument, managementUrl string) map[string]interface{} {
 	return map[string]interface{}{
 		"Id":               vc.Id,
 		"InternalHttpPort": 8080,
@@ -87,8 +77,9 @@ func VChainConfig(vc VChainArgument, genesisValidators []string) map[string]inte
 			"Pull":  false,
 		},
 		"Config": map[string]interface{}{
-			"active-consensus-algo":       2,
-			"genesis-validator-addresses": genesisValidators,
+			"active-consensus-algo": 2,
+			"management-file":       managementUrl,
+			//"genesis-validator-addresses": genesisValidators,
 			//"lean-helix-show-debug":                             true,
 			//"logger-full-log":                                   true,
 
@@ -98,98 +89,56 @@ func VChainConfig(vc VChainArgument, genesisValidators []string) map[string]inte
 	}
 }
 
-type KeyConfig struct {
-	NodeAddress    string `json:"node-address"`
-	NodePrivateKey string `json:"node-private-key,omitempty"` // Very important to omit empty value to produce a valid config
-}
-
-func serveConfig(configStr *string) *httptest.Server {
-	l, err := net.Listen("tcp", fmt.Sprintf("%s:0", helpers.LocalIP()))
-	if err != nil {
-		panic(err)
+func VChainManagementConfig(vc VChainArgument, topology interface{}, genesisValidator []string) map[string]interface{} {
+	var committee []map[string]interface{}
+	for _, validator := range genesisValidator {
+		committee = append(committee, map[string]interface{}{
+			"OrbsAddress":  validator,
+			"Weight":       1000,
+			"IdentityType": 0,
+		})
 	}
 
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		//fmt.Println("configuration requested")
-		_, _ = fmt.Fprint(w, *configStr)
-	})
-
-	server := &httptest.Server{
-		Listener: l,
-		Config: &http.Server{
-			Handler: handler,
+	chains := make(map[string]interface{})
+	chains[fmt.Sprintf("%d", vc.Id)] = map[string]interface{}{
+		"VirtualChainId":  vc.Id,
+		"GenesisRefTime":  0,
+		"CurrentTopology": topology,
+		"CommitteeEvents": []interface{}{
+			map[string]interface{}{
+				"RefTime":   0,
+				"Committee": committee,
+			},
+		},
+		"SubscriptionEvents": []interface{}{
+			map[string]interface{}{
+				"RefTime": 0,
+				"Data": map[string]interface{}{
+					"Status":       "active",
+					"Tier":         "B0",
+					"RolloutGroup": "main",
+					"IdentityType": 0,
+					"Params":       make(map[string]interface{}),
+				},
+			},
+		},
+		"ProtocolVersionEvents": []interface{}{
+			map[string]interface{}{
+				"RefTime": 0,
+				"Data": map[string]interface{}{
+					"RolloutGroup": "main",
+					"Version":      1,
+				},
+			},
 		},
 	}
-	server.Start()
 
-	return server
-}
-
-func SetupDynamicBoyarDependencies(t *testing.T, keyPair KeyConfig, genesisValidators []string, vChains <-chan []VChainArgument) (*config.Flags, func()) {
-	return SetupDynamicBoyarDepencenciesForNetwork(t, keyPair, []interface{}{
-		map[string]interface{}{
-			"address": keyPair.NodeAddress,
-			"ip":      "127.0.0.1",
-		},
-	}, genesisValidators, 80, vChains)
-}
-
-func SetupDynamicBoyarDepencenciesForNetwork(t *testing.T, keyPair KeyConfig,
-	topology []interface{}, genesisValidators []string, httpPort int, vChains <-chan []VChainArgument) (*config.Flags, func()) {
-	return SetupConfigServer(t, keyPair, func() *string {
-		configStr := configJson(t, topology, genesisValidators, httpPort, []VChainArgument{})
-		go func() {
-			for currentChains := range vChains {
-				configStr = configJson(t, topology, genesisValidators, httpPort, currentChains)
-				fmt.Println(configStr)
-			}
-		}()
-
-		return &configStr
-	})
-}
-
-func SetupConfigServer(t *testing.T, keyPair KeyConfig, configBuilder func() *string) (*config.Flags, func()) {
-	keyPairJson, err := json.Marshal(keyPair)
-	require.NoError(t, err)
-	file := TempFile(t, keyPairJson)
-
-	configStr := configBuilder()
-
-	ts := serveConfig(configStr)
-	flags := &config.Flags{
-		Timeout:           time.Minute,
-		ConfigUrl:         ts.URL,
-		KeyPairConfigPath: file.Name(),
-		PollingInterval:   500 * time.Millisecond,
-		WithNamespace:     true,
+	return map[string]interface{}{
+		"CurrentRefTime":   1592834480,
+		"PageStartRefTime": 0,
+		"PageEndRefTime":   1592834480,
+		"VirtualChains":    chains,
 	}
-	cleanup := func() {
-		defer os.Remove(file.Name())
-		defer ts.Close()
-	}
-	return flags, cleanup
-}
-
-func SetupBoyarDependenciesForNetwork(t *testing.T, keyPair KeyConfig, topology []interface{}, genesisValidators []string, httpPort int, vChains ...VChainArgument) (*config.Flags, func()) {
-	vChainsChannel := make(chan []VChainArgument, 1)
-	vChainsChannel <- vChains
-	close(vChainsChannel)
-	return SetupDynamicBoyarDepencenciesForNetwork(t, keyPair, topology, genesisValidators, httpPort, vChainsChannel)
-}
-
-func SetupBoyarDependencies(t *testing.T, keyPair KeyConfig, genesisValidators []string, vChains ...VChainArgument) (*config.Flags, func()) {
-	vChainsChannel := make(chan []VChainArgument, 1)
-	vChainsChannel <- vChains
-	close(vChainsChannel)
-	return SetupDynamicBoyarDependencies(t, keyPair, genesisValidators, vChainsChannel)
-}
-
-func InProcessBoyar(t *testing.T, ctx context.Context, logger log.Logger, flags *config.Flags) govnr.ShutdownWaiter {
-	logger.Info("starting in-process boyar")
-	waiter, err := services.Execute(ctx, flags, logger)
-	require.NoError(t, err)
-	return waiter
 }
 
 func TempFile(t *testing.T, keyPairJson []byte) *os.File {
@@ -223,85 +172,4 @@ func GetVChainMetrics(t helpers.TestingT, port int, vc VChainArgument) JsonMap {
 	err = json.Unmarshal(body, &metrics)
 	require.NoError(t, err)
 	return JsonMap{value: metrics}
-}
-
-func AssertGossipServer(t helpers.TestingT, vc VChainArgument) {
-	timeout := time.Second
-	port := strconv.Itoa(vc.ExternalPort())
-	conn, err := net.DialTimeout("tcp", "127.0.0.1:"+port, timeout)
-	require.NoError(t, err, "error connecting to port %d vChainId %d", port, vc.Id)
-	require.NotNil(t, conn, "nil connection to port %d vChainId %d", port, vc.Id)
-	err = conn.Close()
-	require.NoError(t, err, "closing connection to port %d vChainId %d", port, vc.Id)
-}
-
-func AssertServiceUp(t helpers.TestingT, ctx context.Context, serviceName string) {
-	orchestrator, err := adapter.NewDockerSwarm(adapter.OrchestratorOptions{}, helpers.DefaultTestLogger())
-	require.NoError(t, err)
-
-	statuses, err := orchestrator.GetStatus(ctx, 1*time.Second)
-	require.NoError(t, err)
-
-	ok := false
-	for _, status := range statuses {
-		if status.Name == serviceName && status.State == "started" {
-			ok = true
-			return
-		}
-	}
-
-	require.True(t, ok, "service should be up")
-}
-
-func AssertVolumeExists(t helpers.TestingT, ctx context.Context, volume string) {
-	client := helpers.DockerClient(t)
-
-	res, err := client.VolumeList(ctx, filters.NewArgs(filters.KeyValuePair{
-		Key:   "name",
-		Value: volume,
-	}))
-	require.NoError(t, err)
-
-	require.Len(t, res.Volumes, 1)
-	require.Equal(t, volume, res.Volumes[0].Name)
-}
-
-func AssertServiceDown(t helpers.TestingT, ctx context.Context, serviceName string) {
-	orchestrator, err := adapter.NewDockerSwarm(adapter.OrchestratorOptions{}, helpers.DefaultTestLogger())
-	require.NoError(t, err)
-
-	statuses, err := orchestrator.GetStatus(ctx, 1*time.Second)
-	require.NoError(t, err)
-
-	ok := true
-	for _, status := range statuses {
-		ok = ok && status.Name != serviceName
-	}
-
-	require.True(t, ok, "service should be down")
-}
-
-func AssertVchainUp(t helpers.TestingT, port int, publickKey string, vc1 VChainArgument) {
-	metrics := GetVChainMetrics(t, port, vc1)
-	require.Equal(t, metrics.String("Node.Address"), publickKey)
-	AssertGossipServer(t, vc1)
-}
-
-func AssertVchainDown(t helpers.TestingT, port int, vc1 VChainArgument) {
-	res, err := http.Get(fmt.Sprintf("http://127.0.0.1:%d/vchains/%d/metrics", port, vc1.Id))
-	require.NoError(t, err)
-	require.EqualValues(t, http.StatusNotFound, res.StatusCode)
-}
-
-func AssertManagementServiceUp(t helpers.TestingT, port int) {
-	resp, err := http.Get(fmt.Sprintf("http://127.0.0.1:%d/node/management", port))
-	require.NoError(t, err)
-	require.Equal(t, http.StatusOK, resp.StatusCode)
-	defer resp.Body.Close()
-	body, err := ioutil.ReadAll(resp.Body)
-	require.NoError(t, err)
-
-	config := make(map[string]interface{})
-	err = json.Unmarshal(body, &config)
-	require.NoError(t, err)
 }
